move more internal implementation over to ParticleLoop
