+++ include/neso_particles.hpp
@@ -34,4 +34,5 @@
 #include "utility_mesh_hierarchy_plotting.hpp"
 // Implementations
 #include "departing_particle_identification_impl.hpp"
+#include "global_mapping_impl.hpp"
 #endif
